refactor: split subscriber module
