+++ src/mr.roboto/src/mr/roboto/subscribers/__init__.py
@@ -1,0 +1,7 @@
+from .base import PullRequestSubscriber
+from .contributors_agreement import ContributorsAgreementSigned
+from .explain_trigger_jobs import ExplainHowToTriggerJenkinsJobs
+from .need_to_run_tests import WarnTestsNeedToRun
+from .no_changelog_entry import WarnNoChangelogEntry
+from .trigger_pr_jenkins_jobs import TriggerPullRequestJenkinsJobs
+from .update_coredev import UpdateCoredevCheckouts
