+++ src/mr.roboto/src/mr/roboto/tests/test_add_to_checkouts.py
@@ -1,5 +1,5 @@
 from mr.roboto.events import MergedPullRequest
-from mr.roboto.subscriber import UpdateCoredevCheckouts
+from mr.roboto.subscribers import UpdateCoredevCheckouts
 from tempfile import NamedTemporaryFile
 from unittest import mock
